// src/models/transmit-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * flexCharge-service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdditionalField } from './additional-field';
import { BillingInformation } from './billing-information';
import { Merchant } from './merchant';
import { OrderItem } from './order-item';
import { Payer } from './payer';
import { PaymentMethodLast4Digits } from './payment-method-last4-digits';
import { ShippingInformation } from './shipping-information';
import { Subscription } from './subscription';
import { Transaction } from './transaction';
/**
 *
 * @export
 * @interface TransmitRequest
 */
export interface TransmitRequest {
    /**
     *
     * @type {boolean}
     * @memberof TransmitRequest
     */
    isDeclined: boolean;
    /**
     *
     * @type {string}
     * @memberof TransmitRequest
     */
    mid?: string | null;
    /**
     *
     * @type {string}
     * @memberof TransmitRequest
     */
    orderId: string;
    /**
     *
     * @type {string}
     * @memberof TransmitRequest
     */
    orderSource?: string | null;
    /**
     *
     * @type {Transaction}
     * @memberof TransmitRequest
     */
    transaction: Transaction;
    /**
     *
     * @type {Payer}
     * @memberof TransmitRequest
     */
    payer: Payer;
    /**
     *
     * @type {Array<OrderItem>}
     * @memberof TransmitRequest
     */
    orderItems?: Array<OrderItem> | null;
    /**
     *
     * @type {ShippingInformation}
     * @memberof TransmitRequest
     */
    shippingInformation?: ShippingInformation;
    /**
     *
     * @type {boolean}
     * @memberof TransmitRequest
     */
    isMIT?: boolean | null;
    /**
     *
     * @type {boolean}
     * @memberof TransmitRequest
     */
    isRecurring?: boolean | null;
    /**
     *
     * @type {Subscription}
     * @memberof TransmitRequest
     */
    subscription?: Subscription | null;
    /**
     *
     * @type {Array<AdditionalField>}
     * @memberof TransmitRequest
     */
    additionalFields?: Array<AdditionalField> | null;
    /**
     *
     * @type {Merchant}
     * @memberof TransmitRequest
     */
    merchant?: Merchant | null;
    /**
     *
     * @type {BillingInformation}
     * @memberof TransmitRequest
     */
    billingInformation?: BillingInformation;
    /**
     *
     * @type {PaymentMethodLast4Digits}
     * @memberof TransmitRequest
     */
    paymentMethod?: PaymentMethodLast4Digits;
}